wbwdi: Seamless Access to World Bank World Development Indicators (WDI)

Access and analyze the World Bank’s World Development Indicators (WDI) using the corresponding API https://datahelpdesk.worldbank.org/knowledgebase/articles/889392-about-the-indicators-api-documentation. WDI provides more than 24,000 country or region-level indicators for various contexts. 'wbwdi' enables users to download, process and work with WDI series across multiple countries, aggregates, and time periods.

See also

Author

Maintainer: Christoph Scheuch christoph@tidy-intelligence.com (ORCID) [copyright holder]