wbwdi: Work seamlessly with World Bank World Development Indicators (WDI)

A modern, flexible interface for accessing the World Bank’s World Development Indicators (WDI) API https://datahelpdesk.worldbank.org/knowledgebase/articles/889392-about-the-indicators-api-documentation. 'wbwdi' allows users to download, process, and analyze WDI indicators for multiple geographies and specific time periods. The package is designed to work seamlessly with International Debt Statistics (IDS) provided through the 'wbids' package.

See also

Author

Maintainer: Christoph Scheuch christophh@tidy-intelligence.com (ORCID) [copyright holder]